+++ src/index.js
@@ -2,677 +2,23 @@ if (require('electron-squirrel-startup')) {
     process.exit(0);
 }
 
-const { app, BrowserWindow, desktopCapturer, globalShortcut, session, ipcMain, shell, screen } = require('electron');
-const path = require('node:path');
-const fs = require('node:fs');
-const { GoogleGenAI } = require('@google/genai');
-const os = require('os');
-const { spawn } = require('child_process');
-const { pcmToWav, analyzeAudioBuffer, saveDebugAudio } = require('./audioUtils');
-const { getSystemPrompt } = require('./utils/prompts');
+const { app, BrowserWindow, shell, ipcMain } = require('electron');
+const { createWindow, updateGlobalShortcuts } = require('./utils/window');
+const { setupGeminiIpcHandlers, stopMacOSAudioCapture, sendToRenderer } = require('./utils/gemini');
 
-let geminiSession = null;
-let loopbackProc = null;
-let systemAudioProc = null;
-let audioIntervalTimer = null;
-let mouseEventsIgnored = false;
-let messageBuffer = '';
-let isInitializingSession = false;
+const geminiSessionRef = { current: null };
+let mainWindow = null;
 
-// Add conversation tracking variables
-let currentSessionId = null;
-let currentTranscription = '';
-let conversationHistory = [];
-
-function ensureDataDirectories() {
-    const homeDir = os.homedir();
-    const cheddarDir = path.join(homeDir, 'cheddar');
-    const dataDir = path.join(cheddarDir, 'data');
-    const imageDir = path.join(dataDir, 'image');
-    const audioDir = path.join(dataDir, 'audio');
-
-    [cheddarDir, dataDir, imageDir, audioDir].forEach(dir => {
-        if (!fs.existsSync(dir)) {
-            fs.mkdirSync(dir, { recursive: true });
-        }
-    });
-
-    return { imageDir, audioDir };
+function createMainWindow() {
+    mainWindow = createWindow(sendToRenderer, geminiSessionRef);
+    return mainWindow;
 }
 
-// Conversation management functions
-function initializeNewSession() {
-    currentSessionId = Date.now().toString();
-    currentTranscription = '';
-    conversationHistory = [];
-    console.log('New conversation session started:', currentSessionId);
-}
-
-function saveConversationTurn(transcription, aiResponse) {
-    if (!currentSessionId) {
-        initializeNewSession();
-    }
-
-    const conversationTurn = {
-        timestamp: Date.now(),
-        transcription: transcription.trim(),
-        ai_response: aiResponse.trim(),
-    };
-
-    conversationHistory.push(conversationTurn);
-    console.log('Saved conversation turn:', conversationTurn);
-
-    // Send to renderer to save in IndexedDB
-    sendToRenderer('save-conversation-turn', {
-        sessionId: currentSessionId,
-        turn: conversationTurn,
-        fullHistory: conversationHistory,
-    });
-}
-
-function getCurrentSessionData() {
-    return {
-        sessionId: currentSessionId,
-        history: conversationHistory,
-    };
-}
-
-function createWindow() {
-    // Get layout preference (default to 'normal')
-    let windowWidth = 900;
-    let windowHeight = 400;
-
-    // Try to get layout setting - we'll set up a proper way to read this after window is created
-    // For now, use default normal layout
-
-    const mainWindow = new BrowserWindow({
-        width: windowWidth,
-        height: windowHeight,
-        frame: false,
-        transparent: true,
-        hasShadow: false,
-        alwaysOnTop: true,
-        skipTaskbar: true,
-        hiddenInMissionControl: true,
-        webPreferences: {
-            nodeIntegration: true,
-            contextIsolation: false,
-            backgroundThrottling: false,
-            enableBlinkFeatures: 'GetDisplayMedia',
-            webSecurity: true,
-            allowRunningInsecureContent: false,
-        },
-        backgroundColor: '#00000000',
-    });
-
-    session.defaultSession.setDisplayMediaRequestHandler(
-        (request, callback) => {
-            desktopCapturer.getSources({ types: ['screen'] }).then(sources => {
-                callback({ video: sources[0], audio: 'loopback' });
-            });
-        },
-        { useSystemPicker: true }
-    );
-
-    mainWindow.setResizable(false);
-    mainWindow.setContentProtection(true);
-    mainWindow.setVisibleOnAllWorkspaces(true, { visibleOnFullScreen: true });
-
-    // Center window at the top of the screen
-    const primaryDisplay = screen.getPrimaryDisplay();
-    const { width: screenWidth } = primaryDisplay.workAreaSize;
-    const x = Math.floor((screenWidth - windowWidth) / 2);
-    const y = 0; // Or a small offset like 10, 20 if needed
-    mainWindow.setPosition(x, y);
-
-    if (process.platform === 'win32') {
-        mainWindow.setAlwaysOnTop(true, 'screen-saver', 1);
-    }
-
-    mainWindow.loadFile(path.join(__dirname, 'index.html'));
-
-    // After window is created, check for layout preference and resize if needed
-    mainWindow.webContents.once('dom-ready', () => {
-        // Add a small delay to ensure DOM and localStorage are fully ready
-        setTimeout(() => {
-            // Load keybindings only (layout mode is handled by the UI components)
-            const defaultKeybinds = getDefaultKeybinds();
-            let keybinds = defaultKeybinds;
-
-            // Try to load saved keybinds
-            mainWindow.webContents
-                .executeJavaScript(
-                    `
-                try {
-                    const saved = localStorage.getItem('customKeybinds');
-                    return saved ? JSON.parse(saved) : null;
-                } catch (e) {
-                    return null;
-                }
-            `
-                )
-                .then(savedKeybinds => {
-                    if (savedKeybinds) {
-                        keybinds = { ...defaultKeybinds, ...savedKeybinds };
-                    }
-                    updateGlobalShortcuts(keybinds, mainWindow);
-                })
-                .catch(() => {
-                    // Fallback to default keybinds
-                    updateGlobalShortcuts(keybinds, mainWindow);
-                });
-        }, 150);
-    });
-
-    ipcMain.on('view-changed', (event, view) => {
-        if (view !== 'assistant') {
-            mainWindow.setIgnoreMouseEvents(false);
-        }
-    });
-
-    ipcMain.handle('window-minimize', () => {
-        mainWindow.minimize();
-    });
-
-    // Handle keybind updates
-    ipcMain.on('update-keybinds', (event, newKeybinds) => {
-        updateGlobalShortcuts(newKeybinds, mainWindow);
-    });
-}
-
-function getDefaultKeybinds() {
-    const isMac = process.platform === 'darwin';
-    return {
-        moveUp: isMac ? 'Alt+Up' : 'Ctrl+Up',
-        moveDown: isMac ? 'Alt+Down' : 'Ctrl+Down',
-        moveLeft: isMac ? 'Alt+Left' : 'Ctrl+Left',
-        moveRight: isMac ? 'Alt+Right' : 'Ctrl+Right',
-        toggleVisibility: isMac ? 'Cmd+\\' : 'Ctrl+\\',
-        toggleClickThrough: isMac ? 'Cmd+M' : 'Ctrl+M',
-        nextStep: isMac ? 'Cmd+Enter' : 'Ctrl+Enter',
-        manualScreenshot: isMac ? 'Cmd+Shift+S' : 'Ctrl+Shift+S',
-        previousResponse: isMac ? 'Cmd+[' : 'Ctrl+[',
-        nextResponse: isMac ? 'Cmd+]' : 'Ctrl+]',
-        scrollUp: isMac ? 'Cmd+Shift+Up' : 'Ctrl+Shift+Up',
-        scrollDown: isMac ? 'Cmd+Shift+Down' : 'Ctrl+Shift+Down',
-    };
-}
-
-function updateGlobalShortcuts(keybinds, mainWindow) {
-    console.log('Updating global shortcuts with:', keybinds);
-
-    // Unregister all existing shortcuts
-    globalShortcut.unregisterAll();
-
-    const primaryDisplay = screen.getPrimaryDisplay();
-    const { width, height } = primaryDisplay.workAreaSize;
-    const moveIncrement = Math.floor(Math.min(width, height) * 0.1);
-
-    // Register window movement shortcuts
-    const movementActions = {
-        moveUp: () => {
-            if (!mainWindow.isVisible()) return;
-            const [currentX, currentY] = mainWindow.getPosition();
-            mainWindow.setPosition(currentX, currentY - moveIncrement);
-        },
-        moveDown: () => {
-            if (!mainWindow.isVisible()) return;
-            const [currentX, currentY] = mainWindow.getPosition();
-            mainWindow.setPosition(currentX, currentY + moveIncrement);
-        },
-        moveLeft: () => {
-            if (!mainWindow.isVisible()) return;
-            const [currentX, currentY] = mainWindow.getPosition();
-            mainWindow.setPosition(currentX - moveIncrement, currentY);
-        },
-        moveRight: () => {
-            if (!mainWindow.isVisible()) return;
-            const [currentX, currentY] = mainWindow.getPosition();
-            mainWindow.setPosition(currentX + moveIncrement, currentY);
-        },
-    };
-
-    // Register each movement shortcut
-    Object.keys(movementActions).forEach(action => {
-        const keybind = keybinds[action];
-        if (keybind) {
-            try {
-                globalShortcut.register(keybind, movementActions[action]);
-                console.log(`Registered ${action}: ${keybind}`);
-            } catch (error) {
-                console.error(`Failed to register ${action} (${keybind}):`, error);
-            }
-        }
-    });
-
-    // Register toggle visibility shortcut
-    if (keybinds.toggleVisibility) {
-        try {
-            globalShortcut.register(keybinds.toggleVisibility, () => {
-                if (mainWindow.isVisible()) {
-                    mainWindow.hide();
-                } else {
-                    mainWindow.show();
-                }
-            });
-            console.log(`Registered toggleVisibility: ${keybinds.toggleVisibility}`);
-        } catch (error) {
-            console.error(`Failed to register toggleVisibility (${keybinds.toggleVisibility}):`, error);
-        }
-    }
-
-    // Register toggle click-through shortcut
-    if (keybinds.toggleClickThrough) {
-        try {
-            globalShortcut.register(keybinds.toggleClickThrough, () => {
-                mouseEventsIgnored = !mouseEventsIgnored;
-                if (mouseEventsIgnored) {
-                    mainWindow.setIgnoreMouseEvents(true, { forward: true });
-                    console.log('Mouse events ignored');
-                } else {
-                    mainWindow.setIgnoreMouseEvents(false);
-                    console.log('Mouse events enabled');
-                }
-                mainWindow.webContents.send('click-through-toggled', mouseEventsIgnored);
-            });
-            console.log(`Registered toggleClickThrough: ${keybinds.toggleClickThrough}`);
-        } catch (error) {
-            console.error(`Failed to register toggleClickThrough (${keybinds.toggleClickThrough}):`, error);
-        }
-    }
-
-    // Register next step shortcut
-    if (keybinds.nextStep) {
-        try {
-            globalShortcut.register(keybinds.nextStep, async () => {
-                console.log('Next step shortcut triggered');
-                try {
-                    if (geminiSession) {
-                        await geminiSession.sendRealtimeInput({ text: 'Help me on this page, give me the answer no bs, complete answer' });
-                        console.log('Sent "next step" message to Gemini');
-                    } else {
-                        console.log('No active Gemini session');
-                    }
-                } catch (error) {
-                    console.error('Error sending next step message:', error);
-                }
-            });
-            console.log(`Registered nextStep: ${keybinds.nextStep}`);
-        } catch (error) {
-            console.error(`Failed to register nextStep (${keybinds.nextStep}):`, error);
-        }
-    }
-
-    // Register manual screenshot shortcut
-    if (keybinds.manualScreenshot) {
-        try {
-            globalShortcut.register(keybinds.manualScreenshot, () => {
-                console.log('Manual screenshot shortcut triggered');
-                mainWindow.webContents.executeJavaScript(`
-                    if (window.captureManualScreenshot) {
-                        window.captureManualScreenshot();
-                    } else {
-                        console.log('Manual screenshot function not available');
-                    }
-                `);
-            });
-            console.log(`Registered manualScreenshot: ${keybinds.manualScreenshot}`);
-        } catch (error) {
-            console.error(`Failed to register manualScreenshot (${keybinds.manualScreenshot}):`, error);
-        }
-    }
-
-    // Register previous response shortcut
-    if (keybinds.previousResponse) {
-        try {
-            globalShortcut.register(keybinds.previousResponse, () => {
-                console.log('Previous response shortcut triggered');
-                sendToRenderer('navigate-previous-response');
-            });
-            console.log(`Registered previousResponse: ${keybinds.previousResponse}`);
-        } catch (error) {
-            console.error(`Failed to register previousResponse (${keybinds.previousResponse}):`, error);
-        }
-    }
-
-    // Register next response shortcut
-    if (keybinds.nextResponse) {
-        try {
-            globalShortcut.register(keybinds.nextResponse, () => {
-                console.log('Next response shortcut triggered');
-                sendToRenderer('navigate-next-response');
-            });
-            console.log(`Registered nextResponse: ${keybinds.nextResponse}`);
-        } catch (error) {
-            console.error(`Failed to register nextResponse (${keybinds.nextResponse}):`, error);
-        }
-    }
-
-    // Register scroll up shortcut
-    if (keybinds.scrollUp) {
-        try {
-            globalShortcut.register(keybinds.scrollUp, () => {
-                console.log('Scroll up shortcut triggered');
-                sendToRenderer('scroll-response-up');
-            });
-            console.log(`Registered scrollUp: ${keybinds.scrollUp}`);
-        } catch (error) {
-            console.error(`Failed to register scrollUp (${keybinds.scrollUp}):`, error);
-        }
-    }
-
-    // Register scroll down shortcut
-    if (keybinds.scrollDown) {
-        try {
-            globalShortcut.register(keybinds.scrollDown, () => {
-                console.log('Scroll down shortcut triggered');
-                sendToRenderer('scroll-response-down');
-            });
-            console.log(`Registered scrollDown: ${keybinds.scrollDown}`);
-        } catch (error) {
-            console.error(`Failed to register scrollDown (${keybinds.scrollDown}):`, error);
-        }
-    }
-}
-
-async function initializeGeminiSession(apiKey, customPrompt = '', profile = 'interview', language = 'en-US') {
-    if (isInitializingSession) {
-        console.log('Session initialization already in progress');
-        return false;
-    }
-
-    isInitializingSession = true;
-    sendToRenderer('session-initializing', true);
-
-    const client = new GoogleGenAI({
-        vertexai: false,
-        apiKey: apiKey,
-    });
-
-    // Get enabled tools first to determine Google Search status
-    const enabledTools = await getEnabledTools();
-    const googleSearchEnabled = enabledTools.some(tool => tool.googleSearch);
-
-    const systemPrompt = getSystemPrompt(profile, customPrompt, googleSearchEnabled);
-
-    // Initialize new conversation session
-    initializeNewSession();
-
-    try {
-        const session = await client.live.connect({
-            model: 'gemini-live-2.5-flash-preview',
-            callbacks: {
-                onopen: function () {
-                    sendToRenderer('update-status', 'Connected to Gemini - Starting recording...');
-                },
-                onmessage: function (message) {
-                    console.log('----------------', message);
-
-                    // Handle transcription input
-                    if (message.serverContent?.inputTranscription?.text) {
-                        currentTranscription += message.serverContent.inputTranscription.text;
-                    }
-
-                    // Handle AI model response
-                    if (message.serverContent?.modelTurn?.parts) {
-                        for (const part of message.serverContent.modelTurn.parts) {
-                            console.log(part);
-                            if (part.text) {
-                                messageBuffer += part.text;
-                            }
-                        }
-                    }
-
-                    if (message.serverContent?.generationComplete) {
-                        sendToRenderer('update-response', messageBuffer);
-
-                        // Save conversation turn when we have both transcription and AI response
-                        if (currentTranscription && messageBuffer) {
-                            saveConversationTurn(currentTranscription, messageBuffer);
-                            currentTranscription = ''; // Reset for next turn
-                        }
-
-                        messageBuffer = '';
-                    }
-
-                    if (message.serverContent?.turnComplete) {
-                        sendToRenderer('update-status', 'Listening...');
-                    }
-                },
-                onerror: function (e) {
-                    console.debug('Error:', e.message);
-                    sendToRenderer('update-status', 'Error: ' + e.message);
-                },
-                onclose: function (e) {
-                    console.debug('Session closed:', e.reason);
-                    sendToRenderer('update-status', 'Session closed');
-                },
-            },
-            config: {
-                responseModalities: ['TEXT'],
-                tools: enabledTools,
-                inputAudioTranscription: {},
-                contextWindowCompression: { slidingWindow: {} },
-                speechConfig: { languageCode: language },
-                systemInstruction: {
-                    parts: [{ text: systemPrompt }],
-                },
-            },
-        });
-
-        geminiSession = session;
-        isInitializingSession = false;
-        sendToRenderer('session-initializing', false);
-        return true;
-    } catch (error) {
-        console.error('Failed to initialize Gemini session:', error);
-        isInitializingSession = false;
-        sendToRenderer('session-initializing', false);
-        return false;
-    }
-}
-
-function sendToRenderer(channel, data) {
-    const windows = BrowserWindow.getAllWindows();
-    if (windows.length > 0) {
-        windows[0].webContents.send(channel, data);
-    }
-}
-
-async function getEnabledTools() {
-    const tools = [];
-
-    // Check if Google Search is enabled (default: true)
-    const googleSearchEnabled = await getStoredSetting('googleSearchEnabled', 'true');
-    console.log('Google Search enabled:', googleSearchEnabled);
-
-    if (googleSearchEnabled === 'true') {
-        tools.push({ googleSearch: {} });
-        console.log('Added Google Search tool');
-    } else {
-        console.log('Google Search tool disabled');
-    }
-
-    return tools;
-}
-
-async function getStoredSetting(key, defaultValue) {
-    try {
-        const windows = BrowserWindow.getAllWindows();
-        if (windows.length > 0) {
-            // Wait a bit for the renderer to be ready
-            await new Promise(resolve => setTimeout(resolve, 100));
-
-            // Try to get setting from renderer process localStorage
-            const value = await windows[0].webContents.executeJavaScript(`
-                (function() {
-                    try {
-                        if (typeof localStorage === 'undefined') {
-                            console.log('localStorage not available yet for ${key}');
-                            return '${defaultValue}';
-                        }
-                        const stored = localStorage.getItem('${key}');
-                        console.log('Retrieved setting ${key}:', stored);
-                        return stored || '${defaultValue}';
-                    } catch (e) {
-                        console.error('Error accessing localStorage for ${key}:', e);
-                        return '${defaultValue}';
-                    }
-                })()
-            `);
-            return value;
-        }
-    } catch (error) {
-        console.error('Error getting stored setting for', key, ':', error.message);
-    }
-    console.log('Using default value for', key, ':', defaultValue);
-    return defaultValue;
-}
-
-function killExistingSystemAudioDump() {
-    return new Promise(resolve => {
-        console.log('Checking for existing SystemAudioDump processes...');
-
-        // Kill any existing SystemAudioDump processes
-        const killProc = spawn('pkill', ['-f', 'SystemAudioDump'], {
-            stdio: 'ignore',
-        });
-
-        killProc.on('close', code => {
-            if (code === 0) {
-                console.log('Killed existing SystemAudioDump processes');
-            } else {
-                console.log('No existing SystemAudioDump processes found');
-            }
-            resolve();
-        });
-
-        killProc.on('error', err => {
-            console.log('Error checking for existing processes (this is normal):', err.message);
-            resolve();
-        });
-
-        // Timeout after 2 seconds
-        setTimeout(() => {
-            killProc.kill();
-            resolve();
-        }, 2000);
-    });
-}
-
-async function startMacOSAudioCapture() {
-    if (process.platform !== 'darwin') return false;
-
-    // Kill any existing SystemAudioDump processes first
-    await killExistingSystemAudioDump();
-
-    console.log('Starting macOS audio capture with SystemAudioDump...');
-
-    let systemAudioPath;
-    if (app.isPackaged) {
-        systemAudioPath = path.join(process.resourcesPath, 'SystemAudioDump');
-    } else {
-        systemAudioPath = path.join(__dirname, 'assets', 'SystemAudioDump');
-    }
-
-    console.log('SystemAudioDump path:', systemAudioPath);
-
-    systemAudioProc = spawn(systemAudioPath, [], {
-        stdio: ['ignore', 'pipe', 'pipe'],
-    });
-
-    if (!systemAudioProc.pid) {
-        console.error('Failed to start SystemAudioDump');
-        return false;
-    }
-
-    console.log('SystemAudioDump started with PID:', systemAudioProc.pid);
-
-    const CHUNK_DURATION = 0.1;
-    const SAMPLE_RATE = 24000;
-    const BYTES_PER_SAMPLE = 2;
-    const CHANNELS = 2;
-    const CHUNK_SIZE = SAMPLE_RATE * BYTES_PER_SAMPLE * CHANNELS * CHUNK_DURATION;
-
-    let audioBuffer = Buffer.alloc(0);
-
-    systemAudioProc.stdout.on('data', data => {
-        audioBuffer = Buffer.concat([audioBuffer, data]);
-
-        while (audioBuffer.length >= CHUNK_SIZE) {
-            const chunk = audioBuffer.slice(0, CHUNK_SIZE);
-            audioBuffer = audioBuffer.slice(CHUNK_SIZE);
-
-            const monoChunk = CHANNELS === 2 ? convertStereoToMono(chunk) : chunk;
-            const base64Data = monoChunk.toString('base64');
-            sendAudioToGemini(base64Data);
-
-            if (process.env.DEBUG_AUDIO) {
-                console.log(`Processed audio chunk: ${chunk.length} bytes`);
-                saveDebugAudio(monoChunk, 'system_audio');
-            }
-        }
-
-        const maxBufferSize = SAMPLE_RATE * BYTES_PER_SAMPLE * 1;
-        if (audioBuffer.length > maxBufferSize) {
-            audioBuffer = audioBuffer.slice(-maxBufferSize);
-        }
-    });
-
-    systemAudioProc.stderr.on('data', data => {
-        console.error('SystemAudioDump stderr:', data.toString());
-    });
-
-    systemAudioProc.on('close', code => {
-        console.log('SystemAudioDump process closed with code:', code);
-        systemAudioProc = null;
-    });
-
-    systemAudioProc.on('error', err => {
-        console.error('SystemAudioDump process error:', err);
-        systemAudioProc = null;
-    });
-
-    return true;
-}
-
-function convertStereoToMono(stereoBuffer) {
-    const samples = stereoBuffer.length / 4;
-    const monoBuffer = Buffer.alloc(samples * 2);
-
-    for (let i = 0; i < samples; i++) {
-        const leftSample = stereoBuffer.readInt16LE(i * 4);
-        monoBuffer.writeInt16LE(leftSample, i * 2);
-    }
-
-    return monoBuffer;
-}
-
-function stopMacOSAudioCapture() {
-    if (systemAudioProc) {
-        console.log('Stopping SystemAudioDump...');
-        systemAudioProc.kill('SIGTERM');
-        systemAudioProc = null;
-    }
-}
-
-async function sendAudioToGemini(base64Data) {
-    if (!geminiSession) return;
-
-    try {
-        process.stdout.write('.');
-        await geminiSession.sendRealtimeInput({
-            audio: {
-                data: base64Data,
-                mimeType: 'audio/pcm;rate=24000',
-            },
-        });
-    } catch (error) {
-        console.error('Error sending audio to Gemini:', error);
-    }
-}
-
-app.whenReady().then(createWindow);
+app.whenReady().then(() => {
+    createMainWindow();
+    setupGeminiIpcHandlers(geminiSessionRef);
+    setupGeneralIpcHandlers();
+});
 
 app.on('window-all-closed', () => {
     stopMacOSAudioCapture();
@@ -687,324 +33,35 @@ app.on('before-quit', () => {
 
 app.on('activate', () => {
     if (BrowserWindow.getAllWindows().length === 0) {
-        createWindow();
+        createMainWindow();
     }
 });
 
-ipcMain.handle('initialize-gemini', async (event, apiKey, customPrompt, profile = 'interview', language = 'en-US') => {
-    return await initializeGeminiSession(apiKey, customPrompt, profile, language);
-});
-
-ipcMain.handle('send-audio-content', async (event, { data, mimeType }) => {
-    if (!geminiSession) return { success: false, error: 'No active Gemini session' };
-    try {
-        process.stdout.write('.');
-        await geminiSession.sendRealtimeInput({
-            audio: { data: data, mimeType: mimeType },
-        });
-        return { success: true };
-    } catch (error) {
-        console.error('Error sending audio:', error);
-        return { success: false, error: error.message };
-    }
-});
-
-ipcMain.handle('send-image-content', async (event, { data, debug }) => {
-    if (!geminiSession) return { success: false, error: 'No active Gemini session' };
-
-    try {
-        if (!data || typeof data !== 'string') {
-            console.error('Invalid image data received');
-            return { success: false, error: 'Invalid image data' };
+function setupGeneralIpcHandlers() {
+    ipcMain.handle('quit-application', async event => {
+        try {
+            stopMacOSAudioCapture();
+            app.quit();
+            return { success: true };
+        } catch (error) {
+            console.error('Error quitting application:', error);
+            return { success: false, error: error.message };
         }
+    });
 
-        const buffer = Buffer.from(data, 'base64');
-
-        if (buffer.length < 1000) {
-            console.error(`Image buffer too small: ${buffer.length} bytes`);
-            return { success: false, error: 'Image buffer too small' };
+    ipcMain.handle('open-external', async (event, url) => {
+        try {
+            await shell.openExternal(url);
+            return { success: true };
+        } catch (error) {
+            console.error('Error opening external URL:', error);
+            return { success: false, error: error.message };
         }
+    });
 
-        process.stdout.write('!');
-        await geminiSession.sendRealtimeInput({
-            media: { data: data, mimeType: 'image/jpeg' },
-        });
-
-        return { success: true };
-    } catch (error) {
-        console.error('Error sending image:', error);
-        return { success: false, error: error.message };
-    }
-});
-
-ipcMain.handle('send-text-message', async (event, text) => {
-    if (!geminiSession) return { success: false, error: 'No active Gemini session' };
-
-    try {
-        if (!text || typeof text !== 'string' || text.trim().length === 0) {
-            return { success: false, error: 'Invalid text message' };
+    ipcMain.on('update-keybinds', (event, newKeybinds) => {
+        if (mainWindow) {
+            updateGlobalShortcuts(newKeybinds, mainWindow, sendToRenderer, geminiSessionRef);
         }
-
-        console.log('Sending text message:', text);
-        await geminiSession.sendRealtimeInput({ text: text.trim() });
-        return { success: true };
-    } catch (error) {
-        console.error('Error sending text:', error);
-        return { success: false, error: error.message };
-    }
-});
-
-ipcMain.handle('start-macos-audio', async event => {
-    if (process.platform !== 'darwin') {
-        return {
-            success: false,
-            error: 'macOS audio capture only available on macOS',
-        };
-    }
-
-    try {
-        const success = await startMacOSAudioCapture();
-        return { success };
-    } catch (error) {
-        console.error('Error starting macOS audio capture:', error);
-        return { success: false, error: error.message };
-    }
-});
-
-ipcMain.handle('stop-macos-audio', async event => {
-    try {
-        stopMacOSAudioCapture();
-        return { success: true };
-    } catch (error) {
-        console.error('Error stopping macOS audio capture:', error);
-        return { success: false, error: error.message };
-    }
-});
-
-ipcMain.handle('close-session', async event => {
-    try {
-        stopMacOSAudioCapture();
-
-        // Cleanup any pending resources and stop audio/video capture
-        if (geminiSession) {
-            await geminiSession.close();
-            geminiSession = null;
-        }
-
-        return { success: true };
-    } catch (error) {
-        console.error('Error closing session:', error);
-        return { success: false, error: error.message };
-    }
-});
-
-ipcMain.handle('quit-application', async event => {
-    try {
-        stopMacOSAudioCapture();
-        app.quit();
-        return { success: true };
-    } catch (error) {
-        console.error('Error quitting application:', error);
-        return { success: false, error: error.message };
-    }
-});
-
-ipcMain.handle('open-external', async (event, url) => {
-    try {
-        await shell.openExternal(url);
-        return { success: true };
-    } catch (error) {
-        console.error('Error opening external URL:', error);
-        return { success: false, error: error.message };
-    }
-});
-
-ipcMain.handle('toggle-window-visibility', async event => {
-    try {
-        const windows = BrowserWindow.getAllWindows();
-        if (windows.length > 0) {
-            const mainWindow = windows[0];
-            if (mainWindow.isVisible()) {
-                mainWindow.hide();
-            } else {
-                mainWindow.show();
-            }
-        }
-        return { success: true };
-    } catch (error) {
-        console.error('Error toggling window visibility:', error);
-        return { success: false, error: error.message };
-    }
-});
-
-// Conversation history IPC handlers
-ipcMain.handle('get-current-session', async event => {
-    try {
-        return { success: true, data: getCurrentSessionData() };
-    } catch (error) {
-        console.error('Error getting current session:', error);
-        return { success: false, error: error.message };
-    }
-});
-
-ipcMain.handle('start-new-session', async event => {
-    try {
-        initializeNewSession();
-        return { success: true, sessionId: currentSessionId };
-    } catch (error) {
-        console.error('Error starting new session:', error);
-        return { success: false, error: error.message };
-    }
-});
-
-ipcMain.handle('update-google-search-setting', async (event, enabled) => {
-    try {
-        console.log('Google Search setting updated to:', enabled);
-        // The setting is already saved in localStorage by the renderer
-        // This is just for logging/confirmation
-        return { success: true };
-    } catch (error) {
-        console.error('Error updating Google Search setting:', error);
-        return { success: false, error: error.message };
-    }
-});
-
-// Add layout mode handler
-ipcMain.handle('update-layout-mode', async (event, layoutMode) => {
-    try {
-        console.log('Layout mode update requested:', layoutMode);
-
-        const windows = BrowserWindow.getAllWindows();
-        if (windows.length > 0) {
-            const mainWindow = windows[0];
-
-            let targetWidth, targetHeight;
-
-            if (layoutMode === 'compact') {
-                targetWidth = 700;
-                targetHeight = 300;
-            } else {
-                // Assumes 'normal' or default
-                targetWidth = 900;
-                targetHeight = 400;
-            }
-
-            const [currentWidth, currentHeight] = mainWindow.getSize();
-            console.log('Current window size:', currentWidth, 'x', currentHeight);
-
-            if (currentWidth !== targetWidth || currentHeight !== targetHeight) {
-                // Temporarily enable resizing before changing size
-                mainWindow.setResizable(true);
-                mainWindow.setSize(targetWidth, targetHeight);
-                // Disable resizing after the resize is complete
-                mainWindow.setResizable(false);
-                console.log(`Window resized to ${layoutMode} mode: ${targetWidth}x${targetHeight}`);
-            } else {
-                console.log(`Window already in ${layoutMode} size`);
-            }
-
-            // Re-center the window at the top
-            const primaryDisplay = screen.getPrimaryDisplay();
-            const { width: screenWidth } = primaryDisplay.workAreaSize;
-            const x = Math.floor((screenWidth - targetWidth) / 2);
-            const y = 0; // Position at the top
-            mainWindow.setPosition(x, y);
-            console.log(`Window re-centered to x: ${x}, y: ${y} for ${layoutMode} mode (width: ${targetWidth})`);
-
-            // Verify the resize worked (optional, can be kept or removed)
-            setTimeout(() => {
-                const [newWidth, newHeight] = mainWindow.getSize();
-                console.log('Window size after resize:', newWidth, 'x', newHeight);
-            }, 100);
-        } else {
-            console.warn('No windows found for layout mode update');
-        }
-
-        return { success: true };
-    } catch (error) {
-        console.error('Error updating layout mode:', error);
-        return { success: false, error: error.message };
-    }
-});
-
-// Add view-based window resizing handler
-ipcMain.handle('resize-for-view', async (event, viewName, layoutMode = 'normal') => {
-    try {
-        console.log('View-based resize requested:', viewName, 'layout:', layoutMode);
-
-        const windows = BrowserWindow.getAllWindows();
-        if (windows.length > 0) {
-            const mainWindow = windows[0];
-
-            let targetWidth, targetHeight;
-
-            // Determine base size from layout mode
-            const baseWidth = layoutMode === 'compact' ? 700 : 900;
-            const baseHeight = layoutMode === 'compact' ? 300 : 400;
-
-            // Adjust height based on view
-            switch (viewName) {
-                case 'customize':
-                case 'settings':
-                    targetWidth = baseWidth;
-                    targetHeight = layoutMode === 'compact' ? 500 : 600;
-                    break;
-                case 'help':
-                    targetWidth = baseWidth;
-                    targetHeight = layoutMode === 'compact' ? 450 : 550;
-                    break;
-                case 'history':
-                    targetWidth = baseWidth;
-                    targetHeight = layoutMode === 'compact' ? 450 : 550;
-                    break;
-                case 'advanced':
-                    targetWidth = baseWidth;
-                    targetHeight = layoutMode === 'compact' ? 400 : 500;
-                    break;
-                case 'main':
-                case 'assistant':
-                case 'onboarding':
-                default:
-                    // Use base dimensions for other views
-                    targetWidth = baseWidth;
-                    targetHeight = baseHeight;
-                    break;
-            }
-
-            const [currentWidth, currentHeight] = mainWindow.getSize();
-            console.log('Current window size:', currentWidth, 'x', currentHeight);
-
-            if (currentWidth !== targetWidth || currentHeight !== targetHeight) {
-                // Temporarily enable resizing before changing size
-                mainWindow.setResizable(true);
-                mainWindow.setSize(targetWidth, targetHeight);
-                // Disable resizing after the resize is complete
-                mainWindow.setResizable(false);
-                console.log(`Window resized for ${viewName} view (${layoutMode}): ${targetWidth}x${targetHeight}`);
-            } else {
-                console.log(`Window already correct size for ${viewName} view`);
-            }
-
-            // Re-center the window at the top
-            const primaryDisplay = screen.getPrimaryDisplay();
-            const { width: screenWidth } = primaryDisplay.workAreaSize;
-            const x = Math.floor((screenWidth - targetWidth) / 2);
-            const y = 0; // Position at the top
-            mainWindow.setPosition(x, y);
-            console.log(`Window re-centered to x: ${x}, y: ${y} for ${viewName} view (width: ${targetWidth})`);
-
-            // Verify the resize worked
-            setTimeout(() => {
-                const [newWidth, newHeight] = mainWindow.getSize();
-                console.log('Window size after view resize:', newWidth, 'x', newHeight);
-            }, 50);
-        } else {
-            console.warn('No windows found for view resize');
-        }
-
-        return { success: true };
-    } catch (error) {
-        console.error('Error resizing for view:', error);
-        return { success: false, error: error.message };
-    }
-});
+    });
+}
