+++ src/utils/window.js
@@ -1,0 +1,457 @@
+const { BrowserWindow, globalShortcut, ipcMain, screen } = require('electron');
+const path = require('node:path');
+const fs = require('node:fs');
+const os = require('os');
+
+let mouseEventsIgnored = false;
+
+function ensureDataDirectories() {
+    const homeDir = os.homedir();
+    const cheddarDir = path.join(homeDir, 'cheddar');
+    const dataDir = path.join(cheddarDir, 'data');
+    const imageDir = path.join(dataDir, 'image');
+    const audioDir = path.join(dataDir, 'audio');
+
+    [cheddarDir, dataDir, imageDir, audioDir].forEach(dir => {
+        if (!fs.existsSync(dir)) {
+            fs.mkdirSync(dir, { recursive: true });
+        }
+    });
+
+    return { imageDir, audioDir };
+}
+
+function createWindow(sendToRenderer, geminiSessionRef) {
+    // Get layout preference (default to 'normal')
+    let windowWidth = 900;
+    let windowHeight = 400;
+
+    const mainWindow = new BrowserWindow({
+        width: windowWidth,
+        height: windowHeight,
+        frame: false,
+        transparent: true,
+        hasShadow: false,
+        alwaysOnTop: true,
+        skipTaskbar: true,
+        hiddenInMissionControl: true,
+        webPreferences: {
+            nodeIntegration: true,
+            contextIsolation: false,
+            backgroundThrottling: false,
+            enableBlinkFeatures: 'GetDisplayMedia',
+            webSecurity: true,
+            allowRunningInsecureContent: false,
+        },
+        backgroundColor: '#00000000',
+    });
+
+    const { session, desktopCapturer } = require('electron');
+    session.defaultSession.setDisplayMediaRequestHandler(
+        (request, callback) => {
+            desktopCapturer.getSources({ types: ['screen'] }).then(sources => {
+                callback({ video: sources[0], audio: 'loopback' });
+            });
+        },
+        { useSystemPicker: true }
+    );
+
+    mainWindow.setResizable(false);
+    mainWindow.setContentProtection(true);
+    mainWindow.setVisibleOnAllWorkspaces(true, { visibleOnFullScreen: true });
+
+    // Center window at the top of the screen
+    const primaryDisplay = screen.getPrimaryDisplay();
+    const { width: screenWidth } = primaryDisplay.workAreaSize;
+    const x = Math.floor((screenWidth - windowWidth) / 2);
+    const y = 0;
+    mainWindow.setPosition(x, y);
+
+    if (process.platform === 'win32') {
+        mainWindow.setAlwaysOnTop(true, 'screen-saver', 1);
+    }
+
+    mainWindow.loadFile(path.join(__dirname, '../index.html'));
+
+    // After window is created, check for layout preference and resize if needed
+    mainWindow.webContents.once('dom-ready', () => {
+        setTimeout(() => {
+            const defaultKeybinds = getDefaultKeybinds();
+            let keybinds = defaultKeybinds;
+
+            mainWindow.webContents
+                .executeJavaScript(
+                    `
+                try {
+                    const saved = localStorage.getItem('customKeybinds');
+                    return saved ? JSON.parse(saved) : null;
+                } catch (e) {
+                    return null;
+                }
+            `
+                )
+                .then(savedKeybinds => {
+                    if (savedKeybinds) {
+                        keybinds = { ...defaultKeybinds, ...savedKeybinds };
+                    }
+                    updateGlobalShortcuts(keybinds, mainWindow, sendToRenderer, geminiSessionRef);
+                })
+                .catch(() => {
+                    updateGlobalShortcuts(keybinds, mainWindow, sendToRenderer, geminiSessionRef);
+                });
+        }, 150);
+    });
+
+    setupWindowIpcHandlers(mainWindow, sendToRenderer, geminiSessionRef);
+
+    return mainWindow;
+}
+
+function getDefaultKeybinds() {
+    const isMac = process.platform === 'darwin';
+    return {
+        moveUp: isMac ? 'Alt+Up' : 'Ctrl+Up',
+        moveDown: isMac ? 'Alt+Down' : 'Ctrl+Down',
+        moveLeft: isMac ? 'Alt+Left' : 'Ctrl+Left',
+        moveRight: isMac ? 'Alt+Right' : 'Ctrl+Right',
+        toggleVisibility: isMac ? 'Cmd+\\' : 'Ctrl+\\',
+        toggleClickThrough: isMac ? 'Cmd+M' : 'Ctrl+M',
+        nextStep: isMac ? 'Cmd+Enter' : 'Ctrl+Enter',
+        manualScreenshot: isMac ? 'Cmd+Shift+S' : 'Ctrl+Shift+S',
+        previousResponse: isMac ? 'Cmd+[' : 'Ctrl+[',
+        nextResponse: isMac ? 'Cmd+]' : 'Ctrl+]',
+        scrollUp: isMac ? 'Cmd+Shift+Up' : 'Ctrl+Shift+Up',
+        scrollDown: isMac ? 'Cmd+Shift+Down' : 'Ctrl+Shift+Down',
+    };
+}
+
+function updateGlobalShortcuts(keybinds, mainWindow, sendToRenderer, geminiSessionRef) {
+    console.log('Updating global shortcuts with:', keybinds);
+
+    // Unregister all existing shortcuts
+    globalShortcut.unregisterAll();
+
+    const primaryDisplay = screen.getPrimaryDisplay();
+    const { width, height } = primaryDisplay.workAreaSize;
+    const moveIncrement = Math.floor(Math.min(width, height) * 0.1);
+
+    // Register window movement shortcuts
+    const movementActions = {
+        moveUp: () => {
+            if (!mainWindow.isVisible()) return;
+            const [currentX, currentY] = mainWindow.getPosition();
+            mainWindow.setPosition(currentX, currentY - moveIncrement);
+        },
+        moveDown: () => {
+            if (!mainWindow.isVisible()) return;
+            const [currentX, currentY] = mainWindow.getPosition();
+            mainWindow.setPosition(currentX, currentY + moveIncrement);
+        },
+        moveLeft: () => {
+            if (!mainWindow.isVisible()) return;
+            const [currentX, currentY] = mainWindow.getPosition();
+            mainWindow.setPosition(currentX - moveIncrement, currentY);
+        },
+        moveRight: () => {
+            if (!mainWindow.isVisible()) return;
+            const [currentX, currentY] = mainWindow.getPosition();
+            mainWindow.setPosition(currentX + moveIncrement, currentY);
+        },
+    };
+
+    // Register each movement shortcut
+    Object.keys(movementActions).forEach(action => {
+        const keybind = keybinds[action];
+        if (keybind) {
+            try {
+                globalShortcut.register(keybind, movementActions[action]);
+                console.log(`Registered ${action}: ${keybind}`);
+            } catch (error) {
+                console.error(`Failed to register ${action} (${keybind}):`, error);
+            }
+        }
+    });
+
+    // Register toggle visibility shortcut
+    if (keybinds.toggleVisibility) {
+        try {
+            globalShortcut.register(keybinds.toggleVisibility, () => {
+                if (mainWindow.isVisible()) {
+                    mainWindow.hide();
+                } else {
+                    mainWindow.show();
+                }
+            });
+            console.log(`Registered toggleVisibility: ${keybinds.toggleVisibility}`);
+        } catch (error) {
+            console.error(`Failed to register toggleVisibility (${keybinds.toggleVisibility}):`, error);
+        }
+    }
+
+    // Register toggle click-through shortcut
+    if (keybinds.toggleClickThrough) {
+        try {
+            globalShortcut.register(keybinds.toggleClickThrough, () => {
+                mouseEventsIgnored = !mouseEventsIgnored;
+                if (mouseEventsIgnored) {
+                    mainWindow.setIgnoreMouseEvents(true, { forward: true });
+                    console.log('Mouse events ignored');
+                } else {
+                    mainWindow.setIgnoreMouseEvents(false);
+                    console.log('Mouse events enabled');
+                }
+                mainWindow.webContents.send('click-through-toggled', mouseEventsIgnored);
+            });
+            console.log(`Registered toggleClickThrough: ${keybinds.toggleClickThrough}`);
+        } catch (error) {
+            console.error(`Failed to register toggleClickThrough (${keybinds.toggleClickThrough}):`, error);
+        }
+    }
+
+    // Register next step shortcut
+    if (keybinds.nextStep) {
+        try {
+            globalShortcut.register(keybinds.nextStep, async () => {
+                console.log('Next step shortcut triggered');
+                try {
+                    if (geminiSessionRef.current) {
+                        await geminiSessionRef.current.sendRealtimeInput({ text: 'Help me on this page, give me the answer no bs, complete answer' });
+                        console.log('Sent "next step" message to Gemini');
+                    } else {
+                        console.log('No active Gemini session');
+                    }
+                } catch (error) {
+                    console.error('Error sending next step message:', error);
+                }
+            });
+            console.log(`Registered nextStep: ${keybinds.nextStep}`);
+        } catch (error) {
+            console.error(`Failed to register nextStep (${keybinds.nextStep}):`, error);
+        }
+    }
+
+    // Register manual screenshot shortcut
+    if (keybinds.manualScreenshot) {
+        try {
+            globalShortcut.register(keybinds.manualScreenshot, () => {
+                console.log('Manual screenshot shortcut triggered');
+                mainWindow.webContents.executeJavaScript(`
+                    if (window.captureManualScreenshot) {
+                        window.captureManualScreenshot();
+                    } else {
+                        console.log('Manual screenshot function not available');
+                    }
+                `);
+            });
+            console.log(`Registered manualScreenshot: ${keybinds.manualScreenshot}`);
+        } catch (error) {
+            console.error(`Failed to register manualScreenshot (${keybinds.manualScreenshot}):`, error);
+        }
+    }
+
+    // Register previous response shortcut
+    if (keybinds.previousResponse) {
+        try {
+            globalShortcut.register(keybinds.previousResponse, () => {
+                console.log('Previous response shortcut triggered');
+                sendToRenderer('navigate-previous-response');
+            });
+            console.log(`Registered previousResponse: ${keybinds.previousResponse}`);
+        } catch (error) {
+            console.error(`Failed to register previousResponse (${keybinds.previousResponse}):`, error);
+        }
+    }
+
+    // Register next response shortcut
+    if (keybinds.nextResponse) {
+        try {
+            globalShortcut.register(keybinds.nextResponse, () => {
+                console.log('Next response shortcut triggered');
+                sendToRenderer('navigate-next-response');
+            });
+            console.log(`Registered nextResponse: ${keybinds.nextResponse}`);
+        } catch (error) {
+            console.error(`Failed to register nextResponse (${keybinds.nextResponse}):`, error);
+        }
+    }
+
+    // Register scroll up shortcut
+    if (keybinds.scrollUp) {
+        try {
+            globalShortcut.register(keybinds.scrollUp, () => {
+                console.log('Scroll up shortcut triggered');
+                sendToRenderer('scroll-response-up');
+            });
+            console.log(`Registered scrollUp: ${keybinds.scrollUp}`);
+        } catch (error) {
+            console.error(`Failed to register scrollUp (${keybinds.scrollUp}):`, error);
+        }
+    }
+
+    // Register scroll down shortcut
+    if (keybinds.scrollDown) {
+        try {
+            globalShortcut.register(keybinds.scrollDown, () => {
+                console.log('Scroll down shortcut triggered');
+                sendToRenderer('scroll-response-down');
+            });
+            console.log(`Registered scrollDown: ${keybinds.scrollDown}`);
+        } catch (error) {
+            console.error(`Failed to register scrollDown (${keybinds.scrollDown}):`, error);
+        }
+    }
+}
+
+function setupWindowIpcHandlers(mainWindow, sendToRenderer, geminiSessionRef) {
+    ipcMain.on('view-changed', (event, view) => {
+        if (view !== 'assistant') {
+            mainWindow.setIgnoreMouseEvents(false);
+        }
+    });
+
+    ipcMain.handle('window-minimize', () => {
+        mainWindow.minimize();
+    });
+
+    ipcMain.on('update-keybinds', (event, newKeybinds) => {
+        updateGlobalShortcuts(newKeybinds, mainWindow, sendToRenderer, geminiSessionRef);
+    });
+
+    ipcMain.handle('toggle-window-visibility', async event => {
+        try {
+            if (mainWindow.isVisible()) {
+                mainWindow.hide();
+            } else {
+                mainWindow.show();
+            }
+            return { success: true };
+        } catch (error) {
+            console.error('Error toggling window visibility:', error);
+            return { success: false, error: error.message };
+        }
+    });
+
+    ipcMain.handle('update-layout-mode', async (event, layoutMode) => {
+        try {
+            console.log('Layout mode update requested:', layoutMode);
+
+            let targetWidth, targetHeight;
+
+            if (layoutMode === 'compact') {
+                targetWidth = 700;
+                targetHeight = 300;
+            } else {
+                targetWidth = 900;
+                targetHeight = 400;
+            }
+
+            const [currentWidth, currentHeight] = mainWindow.getSize();
+            console.log('Current window size:', currentWidth, 'x', currentHeight);
+
+            if (currentWidth !== targetWidth || currentHeight !== targetHeight) {
+                mainWindow.setResizable(true);
+                mainWindow.setSize(targetWidth, targetHeight);
+                mainWindow.setResizable(false);
+                console.log(`Window resized to ${layoutMode} mode: ${targetWidth}x${targetHeight}`);
+            } else {
+                console.log(`Window already in ${layoutMode} size`);
+            }
+
+            // Re-center the window at the top
+            const primaryDisplay = screen.getPrimaryDisplay();
+            const { width: screenWidth } = primaryDisplay.workAreaSize;
+            const x = Math.floor((screenWidth - targetWidth) / 2);
+            const y = 0;
+            mainWindow.setPosition(x, y);
+            console.log(`Window re-centered to x: ${x}, y: ${y} for ${layoutMode} mode (width: ${targetWidth})`);
+
+            setTimeout(() => {
+                const [newWidth, newHeight] = mainWindow.getSize();
+                console.log('Window size after resize:', newWidth, 'x', newHeight);
+            }, 100);
+
+            return { success: true };
+        } catch (error) {
+            console.error('Error updating layout mode:', error);
+            return { success: false, error: error.message };
+        }
+    });
+
+    ipcMain.handle('resize-for-view', async (event, viewName, layoutMode = 'normal') => {
+        try {
+            console.log('View-based resize requested:', viewName, 'layout:', layoutMode);
+
+            let targetWidth, targetHeight;
+
+            // Determine base size from layout mode
+            const baseWidth = layoutMode === 'compact' ? 700 : 900;
+            const baseHeight = layoutMode === 'compact' ? 300 : 400;
+
+            // Adjust height based on view
+            switch (viewName) {
+                case 'customize':
+                case 'settings':
+                    targetWidth = baseWidth;
+                    targetHeight = layoutMode === 'compact' ? 500 : 600;
+                    break;
+                case 'help':
+                    targetWidth = baseWidth;
+                    targetHeight = layoutMode === 'compact' ? 450 : 550;
+                    break;
+                case 'history':
+                    targetWidth = baseWidth;
+                    targetHeight = layoutMode === 'compact' ? 450 : 550;
+                    break;
+                case 'advanced':
+                    targetWidth = baseWidth;
+                    targetHeight = layoutMode === 'compact' ? 400 : 500;
+                    break;
+                case 'main':
+                case 'assistant':
+                case 'onboarding':
+                default:
+                    targetWidth = baseWidth;
+                    targetHeight = baseHeight;
+                    break;
+            }
+
+            const [currentWidth, currentHeight] = mainWindow.getSize();
+            console.log('Current window size:', currentWidth, 'x', currentHeight);
+
+            if (currentWidth !== targetWidth || currentHeight !== targetHeight) {
+                mainWindow.setResizable(true);
+                mainWindow.setSize(targetWidth, targetHeight);
+                mainWindow.setResizable(false);
+                console.log(`Window resized for ${viewName} view (${layoutMode}): ${targetWidth}x${targetHeight}`);
+            } else {
+                console.log(`Window already correct size for ${viewName} view`);
+            }
+
+            // Re-center the window at the top
+            const primaryDisplay = screen.getPrimaryDisplay();
+            const { width: screenWidth } = primaryDisplay.workAreaSize;
+            const x = Math.floor((screenWidth - targetWidth) / 2);
+            const y = 0;
+            mainWindow.setPosition(x, y);
+            console.log(`Window re-centered to x: ${x}, y: ${y} for ${viewName} view (width: ${targetWidth})`);
+
+            setTimeout(() => {
+                const [newWidth, newHeight] = mainWindow.getSize();
+                console.log('Window size after view resize:', newWidth, 'x', newHeight);
+            }, 50);
+
+            return { success: true };
+        } catch (error) {
+            console.error('Error resizing for view:', error);
+            return { success: false, error: error.message };
+        }
+    });
+}
+
+module.exports = {
+    ensureDataDirectories,
+    createWindow,
+    getDefaultKeybinds,
+    updateGlobalShortcuts,
+    setupWindowIpcHandlers
+};
